backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
firebase-admin==6.5.0
openai==1.55.3
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.28.1

